Rename basic-auth file
