+++ lib/policies/basic-auth/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  policy: require('./auth'),
+  policy: require('./basic-auth'),
   schema: {
     $id: 'http://express-gateway.io/schemas/policies/basic-auth.json',
     type: 'object',
